fix(reader): 64-bit file sizes/positions — >4GB scan progress was mod-2^32
BaseReader stored the file size in xls::MemBuffer.size and its stream
positions in size_t, which is 32 bits under wasm32. A 4,983,660,763-byte
CSV therefore reported filesize() = 688,693,467 (mod 2^32), so
pos_percent() hit "100%" after the first 656MB of a 4.7GiB scan and the
progress bar parked at 99% for the remaining ~87% of the query (measured
live: byte position "raced" 0->99 in ~5s of a ~60s scan, then sat).
The tail read-back heuristic in underflow() compared against the same
truncated size, and seeks/positions past 4GB would have wrapped too.

Now:
- BaseReader::m_file_size (uint64_t) carries the true size for
  file-backed readers; filesize()/tell()/seek()/do_seek() and both
  stream positions are uint64_t end to end.
- m_content.size stays size_t: it is the read_all() in-memory buffer
  size (ZIP/XLS paths), inherently <4GB in wasm32 — do_open() clamps it
  and read_all() of a >4GB file now fails on allocation instead of
  silently reading into a wrong-sized buffer.

Verified live on the 4.98GB videogames_processed.csv: file-open probe
reports 4752MiB, scan % now climbs monotonically ~2%/s across the whole
~60s COUNT(1) and completes at 98-100% exactly when the query finishes.

+++ extension/datadocs/zip_reader.cpp
@@ -54,8 +54,8 @@ protected:
 		return unzReadCurrentFile(m_zip, buffer, size);
 	}
 
-	virtual bool do_seek(size_t location) override {
-		debug_file_io("ZipReader::do_seek(%zu) NOT_SUPPORTED", location);
+	virtual bool do_seek(uint64_t location) override {
+		debug_file_io("ZipReader::do_seek(%llu) NOT_SUPPORTED", (unsigned long long)location);
 		return false;
 	}
